refactor: CoreDataManager 구현을 통한 Core Data 구현
